+++ AEPLifecycle/Sources/XDM/XDMEnvironment.swift
@@ -13,14 +13,27 @@
 import AEPServices
 import Foundation
 
-/// Represents an XDM Lifecycle Device schema
-struct XDMLifecycleDevice: Encodable {
+/// Represents the Environment schema
+struct XDMEnvironment: Encodable {
     init() {}
 
-    var manufacturer: String?
-    var model: String?
-    var modelNumber: String?
-    var screenHeight: Int64?
-    var screenWidth: Int64?
-    var type: XDMDeviceType?
+    var carrier: String?
+    var language: XDMLifecycleLanguage?
+    var operatingSystem: String?
+    var operatingSystemVersion: String?
+    var type: XDMEnvironmentType?
+
+    enum CodingKeys: String, CodingKey {
+        case carrier
+        case language = "_dc"
+        case operatingSystem
+        case operatingSystemVersion
+        case type
+    }
+
+}
+
+/// Helper struct to encode the language properly in the `XDMLifecycleEnvironment`
+struct XDMLifecycleLanguage: Encodable {
+    let language: String?
 }
