+++ AEPLifecycle/Tests/XDMEnvironmentTests.swift
@@ -9,50 +9,44 @@
 // OF ANY KIND, either express or implied. See the License for the specific language
 // governing permissions and limitations under the License.
 //
+
 @testable import AEPLifecycle
-import AEPServices
-import AEPServicesMocks
 import XCTest
 
-class XDMDeviceTests: XCTestCase {
+class XDMEnvironmentTests: XCTestCase {
 
-    private func buildAndSetMockInfoService() {
-        let mockSystemInfoService = MockSystemInfoService()
-        mockSystemInfoService.deviceName = "test-device-name"
-        mockSystemInfoService.displayInformation = (100, 200)
-        mockSystemInfoService.deviceType = .PHONE
-        ServiceProvider.shared.systemInfoService = mockSystemInfoService
-    }
-
-    // MARK: Encodable Tests
+    // MARK: Encodable tests
 
     func testEncodeEnvironment() throws {
         // setup
-        buildAndSetMockInfoService()
-        var device = XDMDevice()
-        device.model = "test-device-name"
-        device.manufacturer = "apple"
-        device.screenHeight = 200
-        device.screenWidth = 100
-        device.type = .mobile
+        var env = XDMEnvironment()
+        env.operatingSystem = "test-os"
+        env.language = XDMLifecycleLanguage(language: "en-US")
+        env.carrier = "test-carrier"
+        env.operatingSystemVersion = "test-os-version"
+        env.operatingSystem = "test-os-name"
+        env.type = XDMEnvironmentType.application
 
         // test
         let encoder = JSONEncoder()
         encoder.outputFormatting = .prettyPrinted
-        let data = try XCTUnwrap(encoder.encode(device))
+        let data = try XCTUnwrap(encoder.encode(env))
         let dataStr = try XCTUnwrap(String(data: data, encoding: .utf8))
 
         // verify
         let expected = """
         {
-          "manufacturer" : "apple",
-          "model" : "test-device-name",
-          "screenHeight" : 200,
-          "screenWidth" : 100,
-          "type" : "mobile"
+          "operatingSystemVersion" : "test-os-version",
+          "carrier" : "test-carrier",
+          "operatingSystem" : "test-os-name",
+          "type" : "application",
+          "_dc" : {
+            "language" : "en-US"
+          }
         }
         """
 
         XCTAssertEqual(expected, dataStr)
     }
+
 }
