+++ AEPLifecycle/Tests/XDMMobileLifecycleDetailsTests.swift
@@ -21,20 +21,20 @@ class XDMMobileLifecycleDetailsTests: XCTestCase {
         var lifecycleDetails = XDMMobileLifecycleDetails()
         lifecycleDetails.timestamp = Date(timeIntervalSince1970: 1619827190)
         lifecycleDetails.eventType = "test-event-type"
-        lifecycleDetails.application = XDMLifecycleApplication()
+        lifecycleDetails.application = XDMApplication()
         lifecycleDetails.application?.installDate = Date(timeIntervalSince1970: 1619827190)
         lifecycleDetails.application?.isInstall = true
         lifecycleDetails.application?.sessionLength = 424394
         lifecycleDetails.application?.isLaunch = true
         lifecycleDetails.application?.isUpgrade = true
-        lifecycleDetails.environment = XDMLifecycleEnvironment()
+        lifecycleDetails.environment = XDMEnvironment()
         lifecycleDetails.environment?.operatingSystem = "test-os"
         lifecycleDetails.environment?.language = XDMLifecycleLanguage(language: "en-US")
         lifecycleDetails.environment?.carrier = "test-carrier"
         lifecycleDetails.environment?.operatingSystemVersion = "test-os-version"
         lifecycleDetails.environment?.operatingSystem = "test-os-name"
         lifecycleDetails.environment?.type = XDMEnvironmentType.application
-        lifecycleDetails.device = XDMLifecycleDevice()
+        lifecycleDetails.device = XDMDevice()
         lifecycleDetails.device?.model = "test-device-name"
         lifecycleDetails.device?.manufacturer = "apple"
         lifecycleDetails.device?.screenHeight = 200
